가격과 수량 로그에 ToFixedString을 도입하여 더욱 정확한 소수점 정밀도로 로깅하고 과학적 표기법이 등장하지 않도록 수정

+++ Includes/Engines/SymbolInfo.hpp
@@ -59,7 +59,7 @@ class SymbolInfo final {
   [[nodiscard]] double GetPriceStep() const;
   [[nodiscard]] int GetPricePrecision() const;
   [[nodiscard]] double GetQtyStep() const;
-  [[nodiscard]] size_t GetQtyPrecision() const;
+  [[nodiscard]] int GetQtyPrecision() const;
   [[nodiscard]] double GetLimitMaxQty() const;
   [[nodiscard]] double GetLimitMinQty() const;
   [[nodiscard]] double GetMarketMaxQty() const;
